--- test/.gitignore
@@ -1,2 +1,0 @@
-*.csv
-!rawdata.csv
